backend/requirements.txt
fastapi==0.111.1
uvicorn==0.23.1
pypdf
scikit-learn
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
python-dotenv
faiss-cpu
pip install fastapi sqlalchemy mysql-connector-python uvicorn
JSONLoader
TextLoader
Docx2txtLoader
